backend/requirements.txt
Django==4.2.19
django-tenants>=3.6,<4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.7
django-redis==5.4.0
django-celery-beat==2.7.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
pywebpush==2.0.3
python-dotenv==1.0.1
PyYAML==6.0.2
